Dividing up the tests for client
